Fix INT, CALLF, JPF address
Extended address mode use only 16 bit address length before, should be 24 bit.

+++ ana.cpp
@@ -696,7 +696,7 @@ static const opcode_t pre92[] =
 //--------------------------------------------------------------------------
 uint ReadU24(void)
 {
-	ushort value = ua_next_byte() << 16;
+	uint value = ua_next_byte() << 16;
 	value |= ua_next_byte() << 8;
 	value |= ua_next_byte();
 	return value;
